requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
python-dotenv
openpyxl
psycopg2-binary
